Use effect hook for context notification
