+++ renderer/pages/feed.js
@@ -10,8 +10,6 @@ import activeEffect from '../effects/active';
 import scopeOrderMemo from '../memos/scope-order';
 import DropZone from '../components/dropzone';
 
-let contextMessageTimeout;
-
 const Main = () => {
   const [scopes, setScopes] = useState(null);
   const [active, setActive] = useState(null);
@@ -19,7 +17,6 @@ const Main = () => {
   const [config, setConfig] = useState(null);
   const [online, setOnline] = useState(true);
   const [showDropZone, setShowDropZone] = useState(false);
-  const [contextChangedVisible, setContextChanged] = useState(false);
 
   // This effect (and read below)...
   useEffect(() => {
@@ -83,25 +80,10 @@ const Main = () => {
     [JSON.stringify(scopeOrder), JSON.stringify(scopes)]
   );
 
-  const onContextChange = () => {
-    if (!contextMessageTimeout) {
-      setContextChanged(true);
-      contextMessageTimeout = setTimeout(() => {
-        setContextChanged(false);
-        contextMessageTimeout = null;
-      }, 1200);
-    }
-  };
-
   return (
     <main>
       <div onDragEnter={() => setShowDropZone(true)}>
-        <Title
-          config={config}
-          active={active}
-          darkMode={darkMode}
-          contextChanged={contextChangedVisible}
-        />
+        <Title config={config} active={active} darkMode={darkMode} />
 
         {showDropZone && (
           <DropZone darkMode={darkMode} hide={() => setShowDropZone(false)} />
@@ -124,7 +106,6 @@ const Main = () => {
           active={active}
           scopes={orderedScopes}
           setConfig={setConfig}
-          onChange={onContextChange}
         />
       </div>
